Feat S11P12D203-188 : 점령지 정보 조회 타입 수정

+++ backend/mallang/src/main/java/com/chill/mallang/domain/area/service/AreaTopUserService.java
@@ -63,7 +63,7 @@ public class AreaTopUserService {
         for (Answer answer : answers) {
             if (answer.getUser().getFaction() != null && answer.getCheck_fin() == 1) {
                 if (answer.getUser().getFaction().getId() == teamId){
-                    float score = answer.getScore();
+                    double score = answer.getScore();
                     int tryCount = answer.getAnswerTime();
                     teamPoint += (int) score;
                     if (score > maxScore) {
